requirements.txt
tk
pillow
pymunk
